requirements.txt
streamlit==1.42.0
torch==2.6.0
torchvision==0.21.0
pillow==11.1.0
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
PyJWT==2.10.1
opencv-python-headless==4.10.0.84
yt-dlp>=2024.12.13
pypdf==5.1.0
pytest==8.3.4
